refactor: extracts rdf and JSONGraph to separate modules
Two new modules introduced to contain the spesific code to support Rdf and JSONGraph. This reduces
the nunber of dependency that need to be added if you just want the Graph API.

BREAKING CHANGE: Rdf specific code now in @committed/graph-rdf and on the Rdf object. JSONGraph
specific code in @committed/graph-json and on the Json object.

fix #55